HomeWork05-2 Subtask1
